+++ api/src/routes/schedule.ts
@@ -92,11 +92,25 @@ export class ScheduleRouting extends Routing {
 
     @Auth.authorization({ superStudent: true })
     async deleteOne(req: CustomRequest, res: express.Response) {
-        const result = await prisma.schedule.delete({
-            where: {
-                id: Parser.number(req.params["id"]),
-            },
-        });
+        const hardDelete = req.body["hardDelete"];
+        let result;
+
+        if (req.user?.admin && hardDelete) {
+            result = await prisma.schedule.delete({
+                where: {
+                    id: Parser.number(req.params["id"]),
+                },
+            });
+        } else {
+            result = await prisma.schedule.update({
+                data: {
+                    deleted: true,
+                },
+                where: {
+                    id: Parser.number(req.params["id"]),
+                },
+            });
+        }
 
         return res.status(200).json(result);
     }
